Restructure notifications
