+++ app/Http/Controllers/Core/NotificationController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Core;
 
 use App\Http\Controllers\Controller;
-use App\Models\Notification;
+use App\Models\Core\Notification;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
